only fetch branch when needed
