+++ .drone.yml
@@ -1,21 +1,6 @@
 branches:
   exclude: [ gh-pages ]
 
-clone:
-  git:
-    image: plugins/git
-    commands:
-      - git init
-      - git remote add origin ${DRONE_REMOTE_URL}
-      - for remote in `git branch -r`; do git branch --track ${remote#origin/} $remote; done
-      - git fetch --no-tags --all
-      - for b in `git branch`; do git checkout -qf $b; done
-      - if [ "${DRONE_BUILD_EVENT}" == "pull_request" ] || [ "${DRONE_BUILD_EVENT}" == "tag" ] ; then
-          git checkout -qf FETCH_HEAD;
-        else
-          git checkout -qf ${DRONE_COMMIT_BRANCH};
-        fi
-
 matrix:
   TEST:
     - coding-conventions
@@ -72,7 +57,7 @@ pipeline:
     image: node:latest
     secrets: [ SAUCE_USERNAME, SAUCE_ACCESS_KEY ]
     commands:
-      - if [ "${DRONE_BRANCH}" = "master" ] || yarn lerna updated --since master --loglevel silent ; then
+      - if [ "${DRONE_BRANCH}" = "master" ] ; then
           yarn test:functional;
         fi
     when:
