+++ package.json
@@ -38,6 +38,7 @@
     "npmlog": "4.1.2",
     "prettier": "1.11.1",
     "rimraf": "2.6.2",
+    "simple-git": "1.92.0",
     "stylelint": "9.1.1",
     "stylelint-config-prettier": "2.1.0",
     "wdio-mocha-framework": "0.5.12",
