+++ test/utils/specs.js
@@ -1,18 +1,9 @@
 const path = require('path');
 const glob = require('glob');
-const octokit = require('@octokit/rest')();
-
-// Lerna related imports
-const log = require('npmlog');
-const UpdatedPackagesCollector = require('lerna/lib/UpdatedPackagesCollector');
-const Repository = require('lerna/lib/Repository');
-const PackageUtilities = require('lerna/lib/PackageUtilities');
 
 // Utils
 const { isDrone } = require('./drone');
-
-// handle log.success() used by lerna
-log.addLevel('success', 3001, { fg: 'green', bold: true });
+const { getUpdated } = require('../lerna-updated');
 
 module.exports.getSpecs = async () => {
   // By default, test all the specs
@@ -25,40 +16,7 @@ module.exports.getSpecs = async () => {
 
   // Only test the updated components when the branch is not the master
   if (isDrone && process.env.DRONE_BRANCH !== 'master') {
-    log.level = 'silent';
-    const cwd = process.cwd();
-
-    const repo = new Repository(cwd);
-    const packages = PackageUtilities.getPackages(repo);
-    const packageGraph = PackageUtilities.getPackageGraph(packages);
-
-    const matchingPullRequests = await octokit.pullRequests.getAll({
-      owner: process.env.DRONE_REPO_OWNER,
-      repo: process.env.DRONE_REPO_NAME,
-      head: `${process.env.DRONE_REPO_OWNER}:${process.env.DRONE_BRANCH}`,
-    });
-
-    if (!matchingPullRequests || !matchingPullRequests.data) {
-      return specs;
-    }
-
-    const data = matchingPullRequests.data[0];
-
-    const { ref } = data.base;
-
-    console.log('ref', ref);
-
-    // Get updated packages
-    const collector = new UpdatedPackagesCollector({
-      logger: log.newGroup('lerna'),
-      repository: repo,
-      filteredPackages: packages,
-      packageGraph,
-      options: { since: ref },
-      execOpts: { cwd },
-    });
-
-    const updatedPackages = collector.getUpdates();
+    const updatedPackages = getUpdated();
 
     specs = [].concat(
       ...updatedPackages.map(update =>
